backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.58.1
newspaper3k==0.2.8
feedparser==6.0.11
beautifulsoup4==4.12.3
httpx==0.28.1
python-dotenv==1.0.1
lxml==5.3.0
lxml_html_clean==0.4.1
pydantic==2.10.3
sentence-transformers==3.3.1
faiss-cpu==1.9.0
numpy>=1.26.0
sse-starlette==2.1.3
